requirements.txt
playwright==1.40.0
python-dotenv
tqdm
